RAG_ChatBot_python/requirements.txt

langchain>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-text-splitters>=0.3.0,<0.4.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
gradio>=4.0.0,<5.0.0
pypdf2>=3.0.0
pdfplumber>=0.11.0
pypdf>=3.0.0
openai>=1.50.0
tiktoken>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0
torch>=2.0.0
pydantic>=2.0.0,<3.0.0